Fixa un bug nella chiusura della connessione della socket nel server, aggiunge corretta sincronizzazione alla lista dei file nel server da parte delle funzioni in strutturadati.c, aggiorna la documentazione

+++ Server/Utils.c
@@ -81,7 +81,7 @@ void stripCommand(char* input)
 /**
  * @brief Ritorna il numero del processo o del thread che richiama questa funzione.
  */
-int getptid()
+unsigned long int getptid()
 {
 	return procOrThread? getpid() : pthread_self();
 }
