+++ Server/inc/Utils.h
@@ -4,5 +4,5 @@
 void logM (char* messaggio, ...);
 void stripCommand(char*);
 
-int getptid();
+unsigned long int getptid();
 #endif
